Divided the files
Use Start.py to start the game.
